feat(wip): move creators and thunks
