+++ ui/src/notifications/rules/actions/creators.ts
@@ -1,0 +1,44 @@
+// Types
+import {NotificationRuleDraft, Label, RemoteDataState} from 'src/types'
+
+export type Action =
+  | ReturnType<typeof setAllNotificationRules>
+  | ReturnType<typeof setRule>
+  | ReturnType<typeof setCurrentRule>
+  | ReturnType<typeof removeRule>
+  | {
+      type: 'ADD_LABEL_TO_RULE'
+      ruleID: string
+      label: Label
+    }
+  | {
+      type: 'REMOVE_LABEL_FROM_RULE'
+      ruleID: string
+      label: Label
+    }
+
+export const setAllNotificationRules = (
+  status: RemoteDataState,
+  rules?: NotificationRuleDraft[]
+) => ({
+  type: 'SET_ALL_NOTIFICATION_RULES' as 'SET_ALL_NOTIFICATION_RULES',
+  payload: {status, rules},
+})
+
+export const setRule = (rule: NotificationRuleDraft) => ({
+  type: 'SET_NOTIFICATION_RULE' as 'SET_NOTIFICATION_RULE',
+  payload: {rule},
+})
+
+export const setCurrentRule = (
+  status: RemoteDataState,
+  rule?: NotificationRuleDraft
+) => ({
+  type: 'SET_CURRENT_NOTIFICATION_RULE' as 'SET_CURRENT_NOTIFICATION_RULE',
+  payload: {status, rule},
+})
+
+export const removeRule = (ruleID: string) => ({
+  type: 'REMOVE_NOTIFICATION_RULE' as 'REMOVE_NOTIFICATION_RULE',
+  payload: {ruleID},
+})
