+++ ui/src/resources/components/GetResources.tsx
@@ -10,7 +10,7 @@ import {getDashboards} from 'src/dashboards/actions/thunks'
 import {getEndpoints} from 'src/notifications/endpoints/actions/thunks'
 import {getLabels} from 'src/labels/actions'
 import {getMembers} from 'src/members/actions/thunks'
-import {getNotificationRules} from 'src/notifications/rules/actions'
+import {getNotificationRules} from 'src/notifications/rules/actions/thunks'
 import {getPlugins} from 'src/dataLoaders/actions/telegrafEditor'
 import {getScrapers} from 'src/scrapers/actions/thunks'
 import {getTasks} from 'src/tasks/actions/thunks'
