+++ ui/src/store/configureStore.ts
@@ -72,6 +72,7 @@ export const rootReducer = combineReducers<ReducerState>({
     endpoints: endpointsReducer,
     members: membersReducer,
     orgs: orgsReducer,
+    rules: rulesReducer,
     scrapers: scrapersReducer,
     tasks: tasksReducer,
     telegrafs: telegrafsReducer,
@@ -81,7 +82,6 @@ export const rootReducer = combineReducers<ReducerState>({
     views: viewsReducer,
   }),
   routing: routerReducer,
-  rules: rulesReducer,
   telegrafEditor: editorReducer,
   telegrafEditorActivePlugins: activePluginsReducer,
   telegrafEditorPlugins: pluginsReducer,
